draft of documentation for semantics
