+++ src/automaton.rs
@@ -29,6 +29,44 @@ pub use omega::{
 mod with_initial;
 pub use with_initial::Initialized;
 
+/// This module defines different types of _semantics_ that are used by the
+/// [`Automaton`] struct for determining the output of a finite or infinite
+/// run. This can either be some arbitrary output in the case of a Moore or
+/// Mealy machine, or it would be a boolean indicating whether the input is
+/// accepted or not in case of an acceptor like a DFA.
+///
+/// Generally, we distinguish between an [`Automaton`] of finite and one of
+/// infinite words. The purpose of a semantic is to determine what to do
+/// with the run that is induced by an input on some transition system.
+///
+/// # Finite inputs
+/// For a finite input such as a [`FiniteWord`] we use [`FiniteSemantics`],
+/// which defines an `Output` type and provides the [`FininiteSemantics::finite_semantic`] method.
+/// It takes the [`FiniteRun`] that is the result of running the finite
+/// input in some transition system and turns it into the desired output.
+///
+/// Examples of this semantic are for example the [`MooreSemantics`], which
+/// for a finite word `w` simply produce the color of the state that is
+/// reached when running `w` in the transition system from the initial state.
+/// This is similar to the [`DFASemantics`], which additionaly demand that
+/// the state colors are `bool`.
+/// Further, there is also the [`MealySemantics`], which outputs the last
+/// transition that is taken when reading `w`.
+///
+/// # Infinite inputs
+/// For an infinite input like an [`OmegaWord`], we also need to define an
+/// `Output` type. This is now computed in the [`omega_semantic`] method on
+/// the [`OmegaSemantics`] trait. It does this based on an [`OmegaRun`].
+///
+/// Examples include the [`DBASemantics`], which may be applied to `bool`
+/// edge-colored transition systems. It outputs `true` if any edge labeled
+/// with `true` is visited infinitely often and `false` otherwise.
+/// This can actually be viewed as an instantiation of the [`MinEven`]
+/// semantics that a [`DPA`] uses, which outputs the least priority/color
+/// among those that are labels of edges taken infinitely often.
+pub mod semantics;
+pub use semantics::{FiniteSemantics, OmegaSemantics};
+
 /// An automaton consists of a transition system and an acceptance condition.
 /// There are many different types of automata, which can be instantiated from
 /// this struct by setting the type parameters accordingly.
@@ -41,7 +79,7 @@ pub use with_initial::Initialized;
 /// type of the acceptance condition.
 ///
 /// In order for the automaton to be able to accept words, the acceptance condition
-/// must implement the `FiniteSemantics` or `OmegaSemantics` trait, depending on
+/// must implement the [`FiniteSemantics`] or [`OmegaSemantics`] trait, depending on
 /// the value of `OMEGA` (in the former case `OMEGA` should be false, and in the
 /// latter case `OMEGA` should be true).
 #[derive(Clone, Eq, PartialEq, Copy)]
@@ -280,26 +318,6 @@ where
             self.ts
         )
     }
-}
-
-/// This trait is implemented by acceptance conditions for finite words.
-pub trait FiniteSemantics<Q, C> {
-    /// The type of output that this semantic produces.
-    type Output;
-    /// Compute the output for the given finite run.
-    fn finite_semantic<R>(&self, run: R) -> Self::Output
-    where
-        R: FiniteRun<StateColor = Q, EdgeColor = C>;
-}
-
-/// This trait is implemented by acceptance conditions for omega words.
-pub trait OmegaSemantics<Q, C> {
-    /// The type of output that this semantic produces.
-    type Output;
-    /// Compute the output for the given omega run.
-    fn omega_semantic<R>(&self, run: R) -> Self::Output
-    where
-        R: OmegaRun<StateColor = Q, EdgeColor = C>;
 }
 
 /// Iterator over the accepting states of a [`TransitionSystem`] that have a certain coloring.
